mybiz_backend/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
Django==4.2.16
django-allauth==65.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
oauthlib==3.2.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.1
typing-extensions==4.12.2
urllib3==2.2.3
